fix names that didn't get auto-refactored
